Extract SiteCapabilityChecker with caching for capability checks
Move edit_theme_options capability check from SiteListItemBuilder to a new
SiteCapabilityChecker singleton class. Results are cached per site ID to
avoid redundant network requests when rebuilding site list items.
